Improved the performance of love.window.getPixelScale / toPixels / fromPixels by caching some values. Made some coordinate space-sensitive code use the proper conversions to get to the correct coordinate space.
--HG--
branch : minor

+++ src/modules/keyboard/Keyboard.h
@@ -573,7 +573,7 @@ public:
 	 * on the screen the text will appear. This is used as a hint so on-screen
 	 * keyboards don't cover the text area.
 	 **/
-	virtual void setTextInput(bool enable, int x, int y, int w, int h) = 0;
+	virtual void setTextInput(bool enable, double x, double y, double w, double h) = 0;
 
 	/**
 	 * Gets whether text input events are enabled.
